testing int type

+++ src/tinyexpr/tinyexpr.c
@@ -582,12 +582,12 @@ te_value te_interp(const char *expression, int *error) {
     return ret;
 }
 
-static void pn (const te_expr *n, int depth) {
+static void pn(const te_expr *n, int depth) {
     int i, arity;
     printf("%*s", depth, "");
 
     switch(TYPE_MASK(n->type)) {
-    case TE_CONSTANT: printf("%f\n", n->value); break;
+    case TE_CONSTANT: printf("%f\n", (double)n->value); break;
     case TE_VARIABLE: printf("bound %p\n", n->bound); break;
 
     case TE_FUNCTION0: case TE_FUNCTION1: case TE_FUNCTION2: case TE_FUNCTION3:
